talk-yt/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==4.4.2
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
Pillow==9.5.0
proglog==0.1.10
pytube==15.0.0
requests==2.30.0
tqdm==4.65.0
urllib3==2.0.2
websockets==11.0.3
whispercpp==0.0.17
yarl==1.9.2
